src/main/java/com/hladchenko/leetcode/easy/Task1389.java
/*
    Given two arrays of integers nums and index. Your task is to create target array under the following rules:

    Initially target array is empty.
    From left to right read nums[i] and index[i], insert at index index[i] the value nums[i] in target array.
    Repeat the previous step until there are no elements to read in nums and index.
    Return the target array.

    It is guaranteed that the insertion operations will be valid.
    https://leetcode.com/problems/create-target-array-in-the-given-order/
 */
package com.hladchenko.leetcode.easy;

import java.util.ArrayList;
import java.util.List;

public class Task1389 {
    public int[] createTargetArray(int[] nums, int[] index) {
        List<Integer> list = new ArrayList<>();
        for (int i = 0; i < nums.length; i++) {
            list.add(index[i], nums[i]);
        }
        return list.stream().mapToInt(x -> x).toArray();
    }
}
